MDL-82427, fix for class name deprecated
